moved test to different files
